Fix VkResult logging for stricter fmt v12 enum formatting
spdlog v1.17.0 bundles fmt v12, which no longer implicitly formats
arbitrary enums like VkResult. Cast to int in VK_CHECK_RESULT so the
build keeps working with the upgraded submodule.

+++ FlingEngine/Graphics/inc/GraphicsHelpers.h
@@ -11,7 +11,7 @@
 	VkResult res = (f);																\
 	if (res != VK_SUCCESS)															\
 	{																				\
-		F_LOG_ERROR("VkResult is {} in {} at line {}", res, __FILE__, __LINE__);	\
+		F_LOG_ERROR("VkResult is {} in {} at line {}", static_cast<int>(res), __FILE__, __LINE__);	\
 		assert(res == VK_SUCCESS);													\
 	}																				\
 }
